+++ src/python/generator.c
@@ -53,7 +53,7 @@ char* generate(ast_node *node) {
 char* generate_program(ast_node *node) {
 
   char body[1024];
-  int offset = 0;
+  size_t offset = 0;
   int  i;
 
   for(i = 0; i < node->body_length; i++) {
@@ -94,7 +94,7 @@ char* generate_statement(ast_node *node) {
   int output_len = strlen(body) + 3;
 
   char *output = malloc(output_len);
-  int offset = 0;
+  size_t offset = 0;
 
   append(output, &offset, body);
   output[offset++] = '\n';
@@ -128,7 +128,7 @@ char* generate_call (ast_node *node) {
   }
 
   char *output = malloc(name_len + param1_len + param2_len + 5);
-  int offset = 0;
+  size_t offset = 0;
 
   append(output, &offset, node->string_val);
   output[offset++] = '(';
@@ -187,7 +187,7 @@ char* generate_operator (ast_node *node) {
   int param2_len = strlen(param2);
 
   char *output = malloc(param1_len + param2_len + 6);
-  int offset = 0;
+  size_t offset = 0;
 
   output[offset++] = '(';
   append(output, &offset, param1);
