+++ src/util.c
@@ -2,8 +2,8 @@
 
 #include "util.h"
 
-int append(char *dest, int *offset, char const *source) {
-  int size = strlen(source);
+size_t append(char *dest, size_t *offset, char const *source) {
+  size_t size = strlen(source);
   memcpy(&dest[*offset], source, size);
   (*offset) += size;
   return size;
